LemonUI, GTAV 3095 Update
This updates the SHVDN 3.6.0 version of FusionLibrary to better support GTAV v3095, and also fixes several minor bugs.

+++ Replica/MomentReplica.cs
@@ -25,7 +25,7 @@ namespace FusionLibrary
 
         public bool Applied = false;
 
-        public double MomentDuration { get; set; } = 10;
+        public int MomentDuration { get; set; } = 10;
 
         public MomentReplica(DateTime dateTime)
         {
